+++ Foogle-WPF/Models/UserSkills.cs
@@ -8,7 +8,7 @@ namespace Foogle_WPF
 {
     public class UserSkills
     {
-        public int id { get; set; }
+        public Int64 id { get; set; }
 
         public virtual Skill skill { get; set; }
         public virtual FoogleUser user { get; set; }
